DSP-1438 Back to top: focus behaviour when hidden
Fixes the case where the back to top button is hidden in a browser window if the window is not tall enough for it to be useful (i.e. the page top is still visible) but the link can still receive focus while hidden. The link now becomes visible on focus.

DSP-1438 better display when page content is very short